Move saving of measurements into sensor module
